+++ packs/github/actions/lib/base.py
@@ -1,6 +1,7 @@
 from github import Github
 import requests
 from bs4 import BeautifulSoup
+import json
 
 from st2actions.runners.pythonrunner import Action
 
@@ -20,9 +21,13 @@ class BaseGithubAction(Action):
     def __init__(self, config):
         super(BaseGithubAction, self).__init__(config=config)
         token = self.config.get('token', None)
-        token = token or None
-        base_url = self.config.get('base_url', DEFAULT_API_URL)
-        self._client = Github(token, base_url=base_url)
+        self.token = token or None
+        self.github_url = self.config.get('github_url', DEFAULT_API_URL)
+        self.enterprise_url = self.config.get('enterprise_url', None)
+        self.default_github_type = self.config.get('github_type', None)
+
+        self._client = Github(self.token, base_url=self.github_url)
+        self._session = requests.Session()
 
     def _web_session(self):
         '''Returns a requests session to scrape off the web'''
@@ -50,3 +55,61 @@ class BaseGithubAction(Action):
         s = self._web_session()
         response = s.get(url)
         return response.json()
+
+    def _is_enterprise(self, github_type):
+
+        if github_type == "enterprise":
+            return True
+        elif github_type == "online":
+            return False
+        elif self.default_github_type == "enterprise":
+            return True
+        elif self.default_github_type == "online":
+            return False
+        else:
+            raise ValueError("Default GitHub Invalid!")
+
+    def _get_user_token(self, user, enterprise):
+        if enterprise:
+            token_name = "token_enterprise_{}".format(user)
+        else:
+            token_name = "token_{}".format(user)
+
+        return self.action_service.get_value(token_name)
+
+    def _change_to_user_token(self, user, enterprise=False):
+        token = self._get_user_token(user, enterprise)
+
+        if enterprise:
+            self._client = Github(token, base_url=self.enterprise_url)
+        else:
+            self._client = Github(token, base_url=self.github_url)
+
+        return True
+
+    def _request(self, method, uri, payload, token, enterprise):
+        headers = {'Authorization': 'token {}'.format(token)}
+
+        if enterprise:
+            url = "{}{}".format(self.enterprise_url, uri)
+        else:
+            url = "{}{}".format(self.github_url, uri)
+
+        try:
+            r = self._session.request(method,
+                                      url,
+                                      data=json.dumps(payload),
+                                      headers=headers,
+                                      verify=False)
+            r.raise_for_status()
+        except requests.exceptions.HTTPError:
+            raise Exception(
+                "ERROR: '{}'ing to '{}' - status code: {} payload: {}".format(
+                    method, url, r.status_code, json.dumps(payload)))
+        except requests.exceptions.ConnectionError, e:
+            raise Exception("Could not connect to: {} : {}".format(url, e))
+        else:
+            if r.status_code == 204:
+                return None
+            else:
+                return r.json()